dec 29 fix

+++ lib/models/order_review_model.dart
@@ -13,7 +13,7 @@ class OrderReviewModel {
   String? productCode;
   int? paymentMode;
   String? tax;
-  int? offer;
+  double? offer;
   String? billingAddress;
   String? shippingAddress;
   String? netTotal;
